refactor evaluation route to settings route module
