+++ backend/app.js
@@ -11,7 +11,6 @@ const purchaseRouter = require('./routes/purchase');
 const pageURLRouter = require('./routes/webpageURL');
 const settingsRouter = require('./routes/settings');
 const devopsRouter = require('./routes/devops');
-const evaluationRouter = require('./routes/evaluation');
 const path = require('path');
 const filesystem = require('fs');
 
@@ -29,7 +28,6 @@ app.use(purchaseRouter);
 app.use(pageURLRouter);
 app.use(settingsRouter);
 app.use(devopsRouter);
-app.use(evaluationRouter);
 
 const assetsPath = path.join(__dirname, 'assets');
 app.use(express.static(assetsPath, { index: false }));
